move box plot example to a separate file
