+++ packages/demo/examples/01-xy-chart/index.jsx
@@ -32,6 +32,12 @@ import RectPointComponent from './RectPointComponent';
 import ResponsiveXYChart, { dateFormatter } from './ResponsiveXYChart';
 import StackedAreaExample from './StackedAreaExample';
 import ScatterWithHistogram from './ScatterWithHistograms';
+import {
+  SimpleBoxPlotSeriesExample,
+  SingleBoxPlotSeriesExample,
+  HorizontalBoxPlotViolinPlotSeriesExample,
+  BoxPlotViolinPlotSeriesExample,
+} from './StatsSeriesExample';
 
 import {
   circlePackData,
@@ -45,7 +51,6 @@ import {
   intervalData,
   temperatureBands,
   priceBandData,
-  statsData,
 } from './data';
 
 import WithToggle from '../shared/WithToggle';
@@ -494,214 +499,30 @@ export default {
     {
       description: 'Box Plot Example',
       components: [BoxPlotSeries],
-      example: () => {
-        const boxPlotData = statsData.map(s => s.boxPlot);
-        const values = boxPlotData.reduce((r, e) => r.push(e.min, e.max, ...e.outliers) && r, []);
-        const minYValue = Math.min(...values);
-        const maxYValue = Math.max(...values);
-        const yDomain = [minYValue - (0.1 * Math.abs(minYValue)),
-          maxYValue + (0.1 * Math.abs(minYValue))];
-        return (
-          <ResponsiveXYChart
-            theme={{}}
-            ariaLabel="Required label"
-            xScale={{
-              type: 'band',
-              paddingInner: 0.15,
-              paddingOuter: 0.3,
-            }}
-            yScale={{ type: 'linear',
-              domain: yDomain,
-            }}
-            showYGrid
-          >
-            <LinearGradient
-              id="aqua_lightaqua_gradient"
-              from="#99e9f2"
-              to="#c5f6fa"
-            />
-            <YAxis numTicks={4} />
-            <BoxPlotSeries
-              data={boxPlotData}
-              label="Test"
-              fill="url(#aqua_lightaqua_gradient)"
-              stroke="#22b8cf"
-              strokeWidth={1.5}
-            />
-            <XAxis />
-          </ResponsiveXYChart>
-        );
-      },
+      example: () => (
+        <SimpleBoxPlotSeriesExample />
+      ),
     },
     {
       description: 'Single Horizontal Box Plot Example',
       components: [BoxPlotSeries],
-      example: () => {
-        const singleStats = [statsData[0]];
-        const boxPlotData = singleStats.map((s) => {
-          const { boxPlot } = s;
-          const { x, ...rest } = boxPlot;
-          return {
-            y: x,
-            ...rest,
-          };
-        });
-        const values = boxPlotData.reduce((r, e) => r.push(e.min, e.max, ...e.outliers) && r, []);
-        const minXValue = Math.min(...values);
-        const maxXValue = Math.max(...values);
-        const xDomain = [minXValue - (0.1 * Math.abs(minXValue)),
-          maxXValue + (0.1 * Math.abs(maxXValue))];
-        return (
-          <ResponsiveXYChart
-            theme={{}}
-            ariaLabel="Required label"
-            yScale={{
-              type: 'band',
-              paddingInner: 0.15,
-              paddingOuter: 0.3,
-            }}
-            xScale={{ type: 'linear',
-              domain: xDomain,
-            }}
-            showYGrid
-          >
-            <LinearGradient
-              id="aqua_lightaqua_gradient"
-              from="#99e9f2"
-              to="#c5f6fa"
-            />
-            <BoxPlotSeries
-              data={boxPlotData}
-              label="Test"
-              fill="url(#aqua_lightaqua_gradient)"
-              stroke="#22b8cf"
-              strokeWidth={1.5}
-              horizontal
-            />
-            <XAxis />
-          </ResponsiveXYChart>
-        );
-      },
+      example: () => (
+        <SingleBoxPlotSeriesExample />
+      ),
     },
     {
       description: 'Horizontal BoxPlot With ViolinPlot Example',
       components: [BoxPlotSeries, ViolinPlotSeries],
-      example: () => {
-        const boxPlotData = statsData.map((s) => {
-          const { boxPlot } = s;
-          const { x, ...rest } = boxPlot;
-          return {
-            y: x,
-            ...rest,
-          };
-        });
-        const violinData = statsData.map(s => ({ y: s.boxPlot.x, binData: s.binData }));
-        const values = boxPlotData.reduce((r, e) => r.push(e.min, e.max, ...e.outliers) && r, []);
-        const minXValue = Math.min(...values);
-        const maxXValue = Math.max(...values);
-        const xDomain = [minXValue - (0.1 * Math.abs(minXValue)),
-          maxXValue + (0.1 * Math.abs(maxXValue))];
-        return (
-          <ResponsiveXYChart
-            theme={{}}
-            ariaLabel="Required label"
-            yScale={{
-              type: 'band',
-              paddingInner: 0.15,
-              paddingOuter: 0.3,
-            }}
-            xScale={{ type: 'linear',
-              domain: xDomain,
-            }}
-            showYGrid
-          >
-            <PatternLines
-              id="vViolinLines"
-              height={3}
-              width={3}
-              stroke="#ced4da"
-              strokeWidth={1}
-              fill="rgba(0,0,0,0.3)"
-            />
-            <YAxis numTicks={4} />
-            <ViolinPlotSeries
-              data={violinData}
-              label="Test"
-              fill="url(#vViolinLines)"
-              stroke="#22b8cf"
-              strokeWidth={0.5}
-              horizontal
-            />
-            <BoxPlotSeries
-              data={boxPlotData}
-              label="Test"
-              fill={colors.categories[0]}
-              stroke={colors.categories[0]}
-              widthRatio={0.5}
-              fillOpacity={0.2}
-              strokeWidth={1}
-              horizontal
-            />
-            <XAxis />
-          </ResponsiveXYChart>
-        );
-      },
+      example: () => (
+        <HorizontalBoxPlotViolinPlotSeriesExample />
+      ),
     },
     {
       description: 'BoxPlot With ViolinPlot Example',
       components: [BoxPlotSeries, ViolinPlotSeries],
-      example: () => {
-        const boxPlotData = statsData.map(s => s.boxPlot);
-        const violinData = statsData.map(s => ({ x: s.boxPlot.x, binData: s.binData }));
-        const values = boxPlotData.reduce((r, e) => r.push(e.min, e.max, ...e.outliers) && r, []);
-        const minYValue = Math.min(...values);
-        const maxYValue = Math.max(...values);
-        const yDomain = [minYValue - (0.1 * Math.abs(minYValue)),
-          maxYValue + (0.1 * Math.abs(minYValue))];
-        return (
-          <ResponsiveXYChart
-            theme={{}}
-            ariaLabel="Required label"
-            xScale={{
-              type: 'band',
-              paddingInner: 0.15,
-              paddingOuter: 0.3,
-            }}
-            yScale={{ type: 'linear',
-              domain: yDomain,
-            }}
-            showYGrid
-          >
-            <PatternLines
-              id="hViolinLines"
-              height={3}
-              width={3}
-              stroke="#ced4da"
-              strokeWidth={1}
-              fill="rgba(0,0,0,0.3)"
-              orientation={['horizontal']}
-            />
-            <YAxis numTicks={4} />
-            <ViolinPlotSeries
-              data={violinData}
-              label="Test"
-              fill="url(#hViolinLines)"
-              stroke="#22b8cf"
-              strokeWidth={0.5}
-            />
-            <BoxPlotSeries
-              data={boxPlotData}
-              label="Test"
-              fill={colors.categories[0]}
-              stroke={colors.categories[0]}
-              widthRatio={0.5}
-              fillOpacity={0.2}
-              strokeWidth={1}
-            />
-            <XAxis />
-          </ResponsiveXYChart>
-        );
-      },
+      example: () => (
+        <BoxPlotViolinPlotSeriesExample />
+      ),
     },
     {
       description: 'XAxis, YAxis -- orientation',
